requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
numpy>=1.21.0
tqdm>=4.64.0
scikit-learn>=1.1.0
tensorboard>=2.10.0
wandb>=0.13.0
streamlit>=1.25.0
gradio>=3.40.0
Pillow>=9.0.0
opencv-python>=4.6.0
albumentations>=1.3.0
